% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical.R
\name{is.odd}
\alias{is.odd}
\alias{is.even}
\title{Number Type}
\usage{
is.odd(x)

is.even(x)
}
\arguments{
\item{x}{numeric object to be tested}
}
\value{
a vector of type logical
}
\description{
\code{is.odd} tests
whether a number is odd. \code{is.even}
returns the opposite
}
\note{
if \code{x} contains \code{NA} the function returns \code{NA}
}
\examples{
x <- seq(1, 9, 2)
is.odd(x)
y <- seq(2, 8, 2)
is.even(y)
# NA values
x[1] <- NA
is.odd(x)

}
\keyword{math}
